+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-06-10 12:26:25
-// Commit: feat(ui): add bi-directional event bus handler for ui
+// Updated at: 2026-06-11 09:19:15
+// Commit: refactor(ui): extract reusable atomic widgets from ui
 
 class uiComponentState {
-  final String commitRef = '1868';
-  final String lastMessage = 'feat(ui): add bi-directional event bus handler for ui';
-  final DateTime timestamp = DateTime.parse('2026-06-10T12:26:25+05:00');
+  final String commitRef = '1889';
+  final String lastMessage = 'refactor(ui): extract reusable atomic widgets from ui';
+  final DateTime timestamp = DateTime.parse('2026-06-11T09:19:15+05:00');
 }